fix(sdk): format principal arguments
