+++ components/clarinet-sdk-wasm/src/utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod clarity_values;
 pub mod costs;
 pub mod events;
