+++ components/clarinet-sdk/tests/fixtures/contracts/multiplier-contract.clar
@@ -1,0 +1,5 @@
+(impl-trait .multiplier-trait.multiplier)
+
+(define-read-only (multiply (a uint) (b uint))
+  (ok (* a b))
+)
