+++ components/clarinet-sdk/tests/fixtures/contracts/multiplier-trait.clar
@@ -1,0 +1,5 @@
+(define-trait multiplier
+  (
+    (multiply (uint uint) (response uint uint))
+  )
+)
